Migliorata l'implementazione seriale per accettare insiemi di dati più grandi.

+++ serial_impl/serial.h
@@ -2,10 +2,10 @@
 
 #include <iostream>
 #include <ctime>
-#include "../common/data_loading.h"
+#include "../common/data_generation.h"
 
-void merge_sort(int *a, int left, int right);
+void merge_sort(int *a, uint64_t left, uint64_t right);
 
-void merge(int *a, int left, int center, int right);
+void merge(int *a, uint64_t left, uint64_t center, uint64_t right);
 
-void print_array(int *a, int len);
+void print_array(uint64_t *a, uint64_t len);
